Add proper copyright attributions as suggested by FSF.

+++ gnubg/matchequity.c
@@ -1,14 +1,11 @@
 /*
- * matchequity.c
- * by Joern Thyssen, 1999-2002
+ * Copyright (C) 1999-2004 Joern Thyssen <jth@gnubg.org>
+ * Copyright (C) 2000-2016 the AUTHORS
  *
- * Calculate matchequity table based on 
- * formulas from [insert the Managent Science ref here].
- *
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of version 3 or later of the GNU General Public License as
- * published by the Free Software Foundation.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -16,8 +13,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  *
  * $Id$
  */
@@ -715,7 +711,7 @@ ExtendMET(float aarMET[MAXSCORE][MAXSCORE], const int nMaxScore)
 
             rSigma = sqrtf(rStddev0 * rStddev0 + rStddev1 * rStddev1) * sqrtf(rGames);
 
-            if (6.0f * rSigma > nScore0 - nScore1)
+            if (6.0f * rSigma > (float) (nScore0 - nScore1))
                 aarMET[i][j] = NormalDistArea((float) (nScore0 - nScore1), 6.0f * rSigma, 0.0f, rSigma);
             else
                 aarMET[i][j] = 0.0f;
